+++ jacobF.cpp
@@ -1,840 +1,26 @@
-//cs335 Spring 2015
-//
-//program: asteroids.cpp
-//author:  Gordon Griesel
-//date:    2014
-//mod spring 2015: added constructors
-//
-//This program is a game starting point for 335
-//
-// Possible requirements:
-// ----------------------
-// welcome screen
-// menu
-// multiple simultaneous key-press
-// show exhaust for thrusting
-// move the asteroids
-// collision detection for bullet on asteroid
-// collision detection for asteroid on ship
-// control of bullet launch point
-// life span for each bullet
-// cleanup the bullets that miss a target
-// split asteroids into pieces when blasted
-// random generation of new asteroids
-// score keeping
-// levels of difficulty
-// sound
-// use of textures
-//
-
-#include <iostream>
-#include <cstdlib>
-#include <cstring>
-#include <unistd.h>
-#include <ctime>
-#include <cmath>
-#include <X11/Xlib.h>
-//#include <X11/Xutil.h>
-//#include <GL/gl.h>
-//#include <GL/glu.h>
-#include <X11/keysym.h>
-#include <GL/glx.h>
-#include "ppm.h"
-#include "log.h"
 #include "Ball.h"
-#include "fonts.h"
-
-#include "player.cpp"
-#include "keithH.cpp"
-
-
-//defined types
-typedef float Flt;
-typedef float Vec[3];
-typedef Flt	Matrix[4][4];
-
-//macros
-#define rnd() (((double)rand())/(double)RAND_MAX)
-#define random(a) (rand()%a)
-#define VecZero(v) (v)[0]=0.0,(v)[1]=0.0,(v)[2]=0.0
-#define MakeVector(x, y, z, v) (v)[0]=(x),(v)[1]=(y),(v)[2]=(z)
-#define VecCopy(a,b) (b)[0]=(a)[0];(b)[1]=(a)[1];(b)[2]=(a)[2]
-#define VecDot(a,b)	((a)[0]*(b)[0]+(a)[1]*(b)[1]+(a)[2]*(b)[2])
-#define VecSub(a,b,c) (c)[0]=(a)[0]-(b)[0]; \
-                             (c)[1]=(a)[1]-(b)[1]; \
-(c)[2]=(a)[2]-(b)[2]
-//constants
-const float timeslice = 1.0f;
-const float gravity = -0.2f;
-#define PI 3.141592653589793
-#define ALPHA 1
-
-//X Windows variables
-Display *dpy;
-Window win;
-GLXContext glc;
-
-
-//-----------------------------------------------------------------------------
-//Setup timers
-const double physicsRate = 1.0 / 60.0;
-const double oobillion = 1.0 / 1e9;
-extern struct timespec timeStart, timeCurrent;
-extern struct timespec timePause;
-extern double physicsCountdown;
-extern double timeSpan;
-extern double timeDiff(struct timespec *start, struct timespec *end);
-extern void timeCopy(struct timespec *dest, struct timespec *source);
-//-----------------------------------------------------------------------------
-
-int xres=1250, yres=900;
-
-
-
-
-struct Ship {
-    Vec dir;
-    Vec pos;
-    Vec vel;
-    float angle;
-    float color[3];
-    Ship() {
-        VecZero(dir);
-        pos[0] = (Flt)(xres/3);
-        pos[1] = (Flt)(yres/2);
-        pos[2] = 0.0f;
-        VecZero(vel);
-        angle = 0.0;
-        color[0] = 2.0;
-        color[1] = 1.0;
-        color[2] = 1.0;
-    }
-};
-
-
-
-struct Bullet {
-    Vec pos;
-    Vec vel;
-    float color[3];
-    struct timespec time;
-    struct Bullet *prev;
-    struct Bullet *next;
-    Bullet() {
-        prev = NULL;
-        next = NULL;
-    }
-};
-
-struct Asteroid {
-    Vec pos;
-    Vec vel;
-    int nverts;
-    Flt radius;
-    Vec vert[8];
-    float angle;
-    float rotate;
-    float color[3];
-    struct Asteroid *prev;
-    struct Asteroid *next;
-    Asteroid() {
-        prev = NULL;
-        next = NULL;
-    }
-};
-
-
-//temporary instance variables
-Ball ball;
-float ballXPos;
-float ballYPos;
-struct Game {
-    Ship ship;
-    Asteroid *ahead;
-    Bullet *bhead;
-    int nasteroids;
-    int nbullets;
-    struct timespec bulletTimer;
-    struct timespec mouseThrustTimer;
-    bool mouseThrustOn;
-    Game() {
-        ahead = NULL;
-        bhead = NULL;
-        nasteroids = 0;
-        nbullets = 0;
-        mouseThrustOn = false;
-    }
-};
-
-
-int keys[65536];
-
-//function prototypes
-void initXWindows(void);
-void init_opengl(void);
-void cleanupXWindows(void);
-void check_resize(XEvent *e);
-void check_mouse(XEvent *e, Game *game);
-int check_keys(XEvent *e, Game *g);
-void init(Game *g);
-void init_sounds(void);
-void physics(Game *game);
-void render(Game *game);
-void set_mouse_position(int x, int y);
-void show_mouse_cursor(const int onoff);
-
-//KEITHS ADDITION:
-Hud *hud;
-Game game;
-Player p1;
-Player p2;
-time_t timeBegin;
-//-----------------
-
-
-
-int main(void)
-{
-
-    logOpen();
-    initXWindows();
-    init_opengl();
-    Game game;
-    init(&game);
-    srand(time(NULL));
-    clock_gettime(CLOCK_REALTIME, &timePause);
-    clock_gettime(CLOCK_REALTIME, &timeStart);
-    set_mouse_position(100,100);
-
-    //KEITHS ADDITION:
-    hud = new Hud(xres ,yres);
-    timeBegin = time(NULL);
-    //-----------------------------
-
-    //init ball variables
-    ballXPos = xres/2;
-    ballYPos = yres/2;
-
-    int done=0;
-
-    while (!done) {
-        if (timeBegin + 2.0 <= time(NULL)){
-        while (XPending(dpy)) {
-            XEvent e;
-            XNextEvent(dpy, &e);
-            check_resize(&e);
-            //check_mouse(&e, &game);
-            done = check_keys(&e, &game);
-        }
-        clock_gettime(CLOCK_REALTIME, &timeCurrent);
-        timeSpan = timeDiff(&timeStart, &timeCurrent);
-        timeCopy(&timeStart, &timeCurrent);
-        physicsCountdown += timeSpan;
-        while (physicsCountdown >= physicsRate) {
-            physics(&game);
-            physicsCountdown -= physicsRate;
-        }
-        }
-        render(&game);
-        glXSwapBuffers(dpy, win);
-    }
-    cleanupXWindows();
-    cleanup_fonts();
-    logClose();
-
-
-
-
-    return 0;
-}
-
-void cleanupXWindows(void)
-{
-    XDestroyWindow(dpy, win);
-    XCloseDisplay(dpy);
-}
-
-void set_title(void)
-{
-    //Set the window title bar.
-    XMapWindow(dpy, win);
-    XStoreName(dpy, win, "Battle Pong!");
-}
-
-void setup_screen_res(const int w, const int h)
-{
-    xres = w;
-    yres = h;
-}
-
-void initXWindows(void)
-{
-    GLint att[] = { GLX_RGBA, GLX_DEPTH_SIZE, 24, GLX_DOUBLEBUFFER, None };
-    //GLint att[] = { GLX_RGBA, GLX_DEPTH_SIZE, 24, None };
-    XSetWindowAttributes swa;
-    setup_screen_res(xres, yres);
-    dpy = XOpenDisplay(NULL);
-    if (dpy == NULL) {
-        std::cout << "\n\tcannot connect to X server" << std::endl;
-        exit(EXIT_FAILURE);
-    }
-    Window root = DefaultRootWindow(dpy);
-    XVisualInfo *vi = glXChooseVisual(dpy, 0, att);
-    if (vi == NULL) {
-        std::cout << "\n\tno appropriate visual found\n" << std::endl;
-        exit(EXIT_FAILURE);
-    }
-    Colormap cmap = XCreateColormap(dpy, root, vi->visual, AllocNone);
-    swa.colormap = cmap;
-    swa.event_mask = ExposureMask | KeyPressMask | KeyReleaseMask |
-        PointerMotionMask | MotionNotify | ButtonPress | ButtonRelease |
-        StructureNotifyMask | SubstructureNotifyMask;
-    win = XCreateWindow(dpy, root, 0, 0, xres, yres, 0,
-            vi->depth, InputOutput, vi->visual,
-            CWColormap | CWEventMask, &swa);
-    set_title();
-    glc = glXCreateContext(dpy, vi, NULL, GL_TRUE);
-    glXMakeCurrent(dpy, win, glc);
-    show_mouse_cursor(0);
-}
-
-void reshape_window(int width, int height)
-{
-    //window has been resized.
-    setup_screen_res(width, height);
-    //
-    glViewport(0, 0, (GLint)width, (GLint)height);
-    glMatrixMode(GL_PROJECTION); glLoadIdentity();
-    glMatrixMode(GL_MODELVIEW); glLoadIdentity();
-    glOrtho(0, xres, 0, yres, -1, 1);
-    set_title();
-}
-
-void init_opengl(void)
-{
-    //OpenGL initialization
-    glViewport(0, 0, xres, yres);
-    //Initialize matrices
-    glMatrixMode(GL_PROJECTION); glLoadIdentity();
-    glMatrixMode(GL_MODELVIEW); glLoadIdentity();
-    //This sets 2D mode (no perspective)
-    glOrtho(0, xres, 0, yres, -1, 1);
-    //
-    glDisable(GL_LIGHTING);
-    glDisable(GL_DEPTH_TEST);
-    glDisable(GL_FOG);
-    glDisable(GL_CULL_FACE);
-    //
-    //Clear the screen to black
-    glClearColor(0.0, 0.0, 0.0, 1.0);
-    //Do this to allow fonts
-    glEnable(GL_TEXTURE_2D);
-    initialize_fonts();
-}
-
-void check_resize(XEvent *e)
-{
-    //The ConfigureNotify is sent by the
-    //server if the window is resized.
-    if (e->type != ConfigureNotify)
-        return;
-    XConfigureEvent xce = e->xconfigure;
-    if (xce.width != xres || xce.height != yres) {
-        //Window size did change.
-        reshape_window(xce.width, xce.height);
-    }
-}
-
-void init(Game *g) {
-    //build 10 asteroids...
-    for (int j=0; j<10; j++) {
-        Asteroid *a = new Asteroid;
-        a->nverts = 8;
-        a->radius = rnd()*80.0 + 40.0;
-        Flt r2 = a->radius / 2.0;
-        Flt angle = 0.0f;
-        Flt inc = (PI * 2.0) / (Flt)a->nverts;
-        for (int i=0; i<a->nverts; i++) {
-            a->vert[i][0] = sin(angle) * (r2 + rnd() * a->radius);
-            a->vert[i][1] = cos(angle) * (r2 + rnd() * a->radius);
-            angle += inc;
-        }
-        a->pos[0] = (Flt)(rand() % xres);
-        a->pos[1] = (Flt)(rand() % yres);
-        a->pos[2] = 0.0f;
-        a->angle = 0.0;
-        a->rotate = rnd() * 4.0 - 2.0;
-        a->color[0] = 0.8;
-        a->color[1] = 0.8;
-        a->color[2] = 0.7;
-        a->vel[0] = (Flt)(rnd()*2.0-1.0);
-        a->vel[1] = (Flt)(rnd()*2.0-1.0);
-        //add to front of linked list
-        a->next = g->ahead;
-        if (g->ahead != NULL)
-            g->ahead->prev = a;
-        g->ahead = a;
-        g->nasteroids++;
-    }
-    clock_gettime(CLOCK_REALTIME, &g->bulletTimer);
-    memset(keys, 0, 65536);
-}
-
-void normalize(Vec v) {
-    Flt len = v[0]*v[0] + v[1]*v[1];
-    if (len == 0.0f) {
-        v[0] = 1.0;
-        v[1] = 0.0;
-        return;
-    }
-    len = 1.0f / sqrt(len);
-    v[0] *= len;
-    v[1] *= len;
-}
-
-void set_mouse_position(int x, int y) {
-    XWarpPointer(dpy, None, win, 0, 0, 0, 0, x, y);
-}
-
-void show_mouse_cursor(const int onoff)
-{
-    if (onoff) {
-        //this removes our own blank cursor.
-        XUndefineCursor(dpy, win);
-        return;
-    }
-    //vars to make blank cursor
-    Pixmap blank;
-    XColor dummy;
-    char data[1] = {0};
-    Cursor cursor;
-    //make a blank cursor
-    blank = XCreateBitmapFromData (dpy, win, data, 1, 1);
-    if (blank == None)
-        std::cout << "error: out of memory." << std::endl;
-    cursor = XCreatePixmapCursor(dpy, blank, blank, &dummy, &dummy, 0, 0);
-    XFreePixmap(dpy, blank);
-    //this makes you the cursor. then set it using this function
-    XDefineCursor(dpy, win, cursor);
-    //after you do not need the cursor anymore use this function.
-    //it will undo the last change done by XDefineCursor
-    //(thus do only use ONCE XDefineCursor and then XUndefineCursor):
-}
-
-//void check_mouse(XEvent *e, Game *g)
-//{
-
-//}
-
-int check_keys(XEvent *e, Game *g){
-    //keyboard input?
-    static int shift=0;
-    int key = XLookupKeysym(&e->xkey, 0);
-    //Log("key: %i\n", key);
-    if (e->type == KeyRelease) {
-        keys[key]=0;
-        if (key == XK_Shift_L || key == XK_Shift_R)
-            shift=0;
-        if(key == XK_W)
-            g->ship.vel[1] = 0;
-            std::cout << "w was released" << std::endl;
-        return 0;
-    }
-
-    if (e->type == KeyPress) {
-        keys[key]=1;
-        if (key == XK_Shift_L || key == XK_Shift_R) {
-            shift=1;
-            return 0;
-        }
-    }
-    else {
-        return 0;
-    }
-
-    g->ship.vel[1] = 0.0f;
-    g->ship.vel[0] = 0.0f;
-    float shipSpeed = 5.0f;
-    if (shift){}
-    switch(key) {
-        case XK_Escape:
-            return 1;
-        case XK_w:
-        std::cout << "w was pressed" << std::endl;
-        g->ship.vel[1] = shipSpeed;
-            break;
-        case XK_s:
-        std::cout << "s was pressed" << std::endl;
-        g->ship.vel[1] = -shipSpeed;
-            break;
-        case XK_a:
-        g->ship.vel[0] = -shipSpeed;
-            break;
-        case XK_d:
-        g->ship.vel[0] = shipSpeed;
-            break;
-        case XK_minus:
-            break;
-
-    }
-    return 0;
+Ball::Ball(){
 
 }
 
-void deleteBullet(Game *g, Bullet *node)
-{
-    //remove a node from linked list
-    if (node->prev == NULL) {
-        if (node->next == NULL) {
-            g->bhead = NULL;
-        } else {
-            node->next->prev = NULL;
-            g->bhead = node->next;
-        }
-    } else {
-        if (node->next == NULL) {
-            node->prev->next = NULL;
-        } else {
-            node->prev->next = node->next;
-            node->next->prev = node->prev;
-        }
-    }
-    delete node;
-    node = NULL;
-}
-
-void deleteAsteroid(Game *g, Asteroid *node)
-{
-    //remove a node from linked list
-    if (g){}
-    if (node){}
-}
-
-void buildAsteroidFragment(Asteroid *ta, Asteroid *a)
-{
-    //build ta from a
-    ta->nverts = 8;
-    ta->radius = a->radius / 2.0;
-    Flt r2 = ta->radius / 2.0;
-    Flt angle = 0.0f;
-    Flt inc = (PI * 2.0) / (Flt)ta->nverts;
-    for (int i=0; i<ta->nverts; i++) {
-        ta->vert[i][0] = sin(angle) * (r2 + rnd() * ta->radius);
-        ta->vert[i][1] = cos(angle) * (r2 + rnd() * ta->radius);
-        angle += inc;
-    }
-    ta->pos[0] = a->pos[0] + rnd()*10.0-5.0;
-    ta->pos[1] = a->pos[1] + rnd()*10.0-5.0;
-    ta->pos[2] = 0.0f;
-    ta->angle = 0.0;
-    ta->rotate = a->rotate + (rnd() * 4.0 - 2.0);
-    ta->color[0] = 0.8;
-    ta->color[1] = 0.8;
-    ta->color[2] = 0.7;
-    ta->vel[0] = a->vel[0] + (rnd()*2.0-1.0);
-    ta->vel[1] = a->vel[1] + (rnd()*2.0-1.0);
-}
-
-void physics(Game *g)
-{
-    Flt d0,d1,dist;
-    //Update ship position
-    g->ship.pos[0] += g->ship.vel[0];
-    g->ship.pos[1] += g->ship.vel[1];
-    //Check for collision with window edges
-    if (g->ship.pos[0] <= 0.0) {
-        g->ship.pos[0] = 0;
-    }
-    else if (g->ship.pos[0] +10 >= (float)xres/2) {
-        g->ship.pos[0] = ((float)xres/2)-10;
-        g->ship.vel[0] = 0;
-    }
-    else if (g->ship.pos[1] <= 0.0) {
-        g->ship.pos[1] = 0;
-    }
-    else if (g->ship.pos[1] +100 >= (float)yres) {
-        g->ship.pos[1] = (float)yres - 100;
-        g->ship.vel[1] = 0;
-    }
-    //
-    //
-    //Update bullet positions
-    struct timespec bt;
-    clock_gettime(CLOCK_REALTIME, &bt);
-    Bullet *b = g->bhead;
-    while (b) {
-        //How long has bullet been alive?
-        double ts = timeDiff(&b->time, &bt);
-        if (ts > 2.5) {
-            //time to delete the bullet.
-            Bullet *saveb = b->next;
-            deleteBullet(g, b);
-            b = saveb;
-            g->nbullets--;
-            continue;
-        }
-        //move the bullet
-        b->pos[0] += b->vel[0];
-        b->pos[1] += b->vel[1];
-        //Check for collision with window edges
-        if (b->pos[0] < 0.0) {
-            b->pos[0] += (float)xres;
-        }
-        else if (b->pos[0] > (float)xres) {
-            b->pos[0] -= (float)xres;
-        }
-        else if (b->pos[1] < 0.0) {
-            b->pos[1] += (float)yres;
-        }
-        else if (b->pos[1] > (float)yres) {
-            b->pos[1] -= (float)yres;
-        }
-        b = b->next;
-    }
-    //
-    //Update asteroid positions
-    Asteroid *a = g->ahead;
-    while (a) {
-        a->pos[0] += a->vel[0];
-        a->pos[1] += a->vel[1];
-        //Check for collision with window edges
-        if (a->pos[0] < -100.0) {
-            a->pos[0] += (float)xres+200;
-        }
-        else if (a->pos[0] > (float)xres+100) {
-            a->pos[0] -= (float)xres+200;
-        }
-        else if (a->pos[1] < -100.0) {
-            a->pos[1] += (float)yres+200;
-        }
-        else if (a->pos[1] > (float)yres+100) {
-            a->pos[1] -= (float)yres+200;
-        }
-        a->angle += a->rotate;
-        a = a->next;
-    }
-    //
-    //Asteroid collision with bullets?
-    //If collision detected:
-    //     1. delete the bullet
-    //     2. break the asteroid into pieces
-    //        if asteroid small, delete it
-    a = g->ahead;
-    while (a) {
-        //is there a bullet within its radius?
-        Bullet *b = g->bhead;
-        while (b) {
-            d0 = b->pos[0] - a->pos[0];
-            d1 = b->pos[1] - a->pos[1];
-            dist = (d0*d0 + d1*d1);
-            if (dist < (a->radius*a->radius)) {
-                //std::cout << "asteroid hit." << std::endl;
-                //this asteroid is hit.
-                if (a->radius > 20.0) {
-                    //break it into pieces.
-                    Asteroid *ta = a;
-                    buildAsteroidFragment(ta, a);
-                    int r = rand()%10+5;
-                    for (int k=0; k<r; k++) {
-                        //get the next asteroid position in the array
-                        Asteroid *ta = new Asteroid;
-                        buildAsteroidFragment(ta, a);
-                        //add to front of asteroid linked list
-                        ta->next = g->ahead;
-                        if (g->ahead != NULL)
-                            g->ahead->prev = ta;
-                        g->ahead = ta;
-                        g->nasteroids++;
-                    }
-                } else {
-                    a->color[0] = 1.0;
-                    a->color[1] = 0.1;
-                    a->color[2] = 0.1;
-                    //asteroid is too small to break up
-                    //delete the asteroid and bullet
-                    Asteroid *savea = a->next;
-                    deleteAsteroid(g, a);
-                    a = savea;
-                    g->nasteroids--;
-                }
-                //delete the bullet...
-                Bullet *saveb = b->next;
-                deleteBullet(g, b);
-                b = saveb;
-                g->nbullets--;
-                if (a == NULL)
-                    break;
-                continue;
-            }
-            b = b->next;
-        }
-        if (a == NULL)
-            break;
-        a = a->next;
-    }
-    //---------------------------------------------------
-    //check keys pressed now
-    if (keys[XK_Left]) {
-        g->ship.angle += 4.0;
-        if (g->ship.angle >= 360.0f)
-            g->ship.angle -= 360.0f;
-    }
-    if (keys[XK_Right]) {
-        g->ship.angle -= 4.0;
-        if (g->ship.angle < 0.0f)
-            g->ship.angle += 360.0f;
-    }
-    if (keys[XK_Up]) {
-        //apply thrust
-        //convert ship angle to radians
-        Flt rad = ((g->ship.angle+90.0) / 360.0f) * PI * 2.0;
-        //convert angle to a vector
-        Flt xdir = cos(rad);
-        Flt ydir = sin(rad);
-        g->ship.vel[0] += xdir*0.02f;
-        g->ship.vel[1] += ydir*0.02f;
-        Flt speed = sqrt(g->ship.vel[0]*g->ship.vel[0]+
-                                        g->ship.vel[1]*g->ship.vel[1]);
-        if (speed > 10.0f) {
-            speed = 10.0f;
-            normalize(g->ship.vel);
-            g->ship.vel[0] *= speed;
-            g->ship.vel[1] *= speed;
-        }
-    }
-    if (keys[XK_space]) {
-        //a little time between each bullet
-        struct timespec bt;
-        clock_gettime(CLOCK_REALTIME, &bt);
-        double ts = timeDiff(&g->bulletTimer, &bt);
-        if (ts > 0.1) {
-            timeCopy(&g->bulletTimer, &bt);
-            //shoot a bullet...
-            Bullet *b = new Bullet;
-            timeCopy(&b->time, &bt);
-            b->pos[0] = g->ship.pos[0];
-            b->pos[1] = g->ship.pos[1];
-            b->vel[0] = g->ship.vel[0];
-            b->vel[1] = g->ship.vel[1];
-            //convert ship angle to radians
-            Flt rad = ((g->ship.angle+90.0) / 360.0f) * PI * 2.0;
-            //convert angle to a vector
-            Flt xdir = cos(rad);
-            Flt ydir = sin(rad);
-            b->pos[0] += xdir*20.0f;
-            b->pos[1] += ydir*20.0f;
-            b->vel[0] += xdir*6.0f + rnd()*0.1;
-            b->vel[1] += ydir*6.0f + rnd()*0.1;
-            b->color[0] = 1.0f;
-            b->color[1] = 1.0f;
-            b->color[2] = 1.0f;
-            //add to front of bullet linked list
-            b->next = g->bhead;
-            if (g->bhead != NULL)
-                g->bhead->prev = b;
-            g->bhead = b;
-            g->nbullets++;
-        }
-
-         if (keys[XK_j]){
-             g->ship.vel[1] += 3.0f;
-         }
-
-    }
-
-
-    //ball physics
-    if(ballXPos >= xres/3){
-        ballXPos -= 3.0f;
-    }
-    else{
-        ballXPos += 1.0f;
-    }
-
-
-
+Ball::~Ball(){
 
 }
 
-void render(Game *g)
-{    
+void Ball::render(float xPos, float yPos, float radius){
+    int i;
+    int triangleAmount = 20;
 
+    float twicePi = 2.0f * 3.14159265358979f;
 
-    glClear(GL_COLOR_BUFFER_BIT);
-
-    //KEITHS ADDITIONS:------------------
-    if (timeBegin + 2.0 > time(NULL)){
-        //PASS showWelcome the high score:
-        hud->showWelcome(0);
-        hud->is_show_welcome = true;
-        return;
-    }
-    hud->is_show_welcome = false;
-    hud->showScore(p1.getScore(), p2.getScore());
-    hud->showCourtYard();
-    //------------------------------------
-
-
-    //-------------------------------------------------------------------------
-    //Draw the ship
-    glColor3fv(g->ship.color);
-    glPushMatrix();
-    glTranslatef(g->ship.pos[0], g->ship.pos[1], g->ship.pos[2]);
-    glRotatef(g->ship.angle, 0.0f, 0.0f, 1.0f);
-    glRectf(0.0f, 0.0f, 10.0f, 100.0f);
+    glBegin(GL_TRIANGLE_FAN);
+        glVertex2f(xPos, yPos); // center of circle
+        for(i = 0; i <= triangleAmount;i++) {
+            glVertex2f(
+                    xPos + (radius * cos(i *  twicePi / triangleAmount)),
+                yPos + (radius * sin(i * twicePi / triangleAmount))
+            );
+        }
     glEnd();
-    glColor3f(1.0f, 0.0f, 0.0f);
-    glBegin(GL_POINTS);
-    glVertex2f(0.0f, 0.0f);
-    glEnd();
-    glPopMatrix();
-    if (keys[XK_Up] || g->mouseThrustOn) {
-        int i;
-        //draw thrust
-        Flt rad = ((g->ship.angle+90.0) / 360.0f) * PI * 2.0;
-        //convert angle to a vector
-        Flt xdir = cos(rad);
-        Flt ydir = sin(rad);
-        Flt xs,ys,xe,ye,r;
-        glBegin(GL_LINES);
-        for (i=0; i<16; i++) {
-            xs = -xdir * 11.0f + rnd() * 4.0 - 2.0;
-            ys = -ydir * 11.0f + rnd() * 4.0 - 2.0;
-            r = rnd()*40.0+40.0;
-            xe = -xdir * r + rnd() * 18.0 - 9.0;
-            ye = -ydir * r + rnd() * 18.0 - 9.0;
-            glColor3f(rnd()*.3+.7, rnd()*.3+.7, 0);
-            glVertex2f(g->ship.pos[0]+xs,g->ship.pos[1]+ys);
-            glVertex2f(g->ship.pos[0]+xe,g->ship.pos[1]+ye);
-        }
-        glEnd();
-    }
-    //-------------------------------------------------------------------------
-    //Draw the ball
-
-    ball.render(ballXPos, 5.0f, ballYPos);
-
-
-    //-------------------------------------------------------------------------
-    //Draw the bullets
-    {
-        Bullet *b = g->bhead;
-        while (b) {
-            //Log("draw bullet...\n");
-            glColor3f(1.0, 1.0, 1.0);
-            glBegin(GL_POINTS);
-            glVertex2f(b->pos[0],      b->pos[1]);
-            glVertex2f(b->pos[0]-1.0f, b->pos[1]);
-            glVertex2f(b->pos[0]+1.0f, b->pos[1]);
-            glVertex2f(b->pos[0],      b->pos[1]-1.0f);
-            glVertex2f(b->pos[0],      b->pos[1]+1.0f);
-            glColor3f(0.8, 0.8, 0.8);
-            glVertex2f(b->pos[0]-1.0f, b->pos[1]-1.0f);
-            glVertex2f(b->pos[0]-1.0f, b->pos[1]+1.0f);
-            glVertex2f(b->pos[0]+1.0f, b->pos[1]-1.0f);
-            glVertex2f(b->pos[0]+1.0f, b->pos[1]+1.0f);
-            glEnd();
-            b = b->next;
-        }
-    }
 }
+
